Move list bots call to internal client
 * Move ListBots meta call to an internal client
   struct, all remaining groupme.Client methods
   should restrict command implementations to
   interact with a single chat
 * Hide bot ID and access token from public
   groupme.Client interface
