+++ .gitignore
@@ -1,3 +1,4 @@
 *.iml
+*.idea
 *.zip
 spongebob-bot
