+++ go.mod
@@ -5,4 +5,6 @@ go 1.13
 require (
 	github.com/aws/aws-lambda-go v1.13.3
 	github.com/sirupsen/logrus v1.4.2
+	golang.org/x/crypto v0.0.0-20200109152110-61a87790db17 // indirect
+	golang.org/x/net v0.0.0-20191209160850-c0dbc17a3553 // indirect
 )
